requirements.txt
Django>=3.0.0
postgres>=3.0.0
Pillow>=7.1.2
djangorestframework>=3.11.0
django-filter>=2.3.0